+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Correctly register deferred messages sent from authoritative matches.
 - Correctly cancel Lua authoritative match context when match initialization fails.
 - Improve decoding of Steam authentication responses to correctly unwrap payload.
+- Correctly parse Steam Web API errors when authenticating Steam tokens.
 
 ## [2.3.0] - 2018-12-31
 ### Added